+++ lib/ops_manager/api/base.rb
@@ -153,7 +153,10 @@ class OpsManager
         Net::HTTP.new(uri.host, uri.port).tap do |http|
           http.use_ssl = true
           http.verify_mode = OpenSSL::SSL::VERIFY_NONE
-          http.read_timeout = 1200
+          http.read_timeout = 1800
+
+          ctx = OpenSSL::SSL::SSLContext.new
+          ctx.ssl_version = :TLSv1_2
         end
       end