+++ spec/ops_manager/logging_spec.rb
@@ -19,13 +19,8 @@ describe OpsManager::Logging do
       expect(foo.logger).to eq(foo.logger)
     end
 
-    it 'should output logs to stdout' do
-      expect(Logger).to receive(:new).with(STDOUT).and_call_original
-      foo.logger
-    end
-
-    it 'should log in WARN level' do
-      expect_any_instance_of(Logger).to receive(:level=).with(Logger::WARN)
+    it 'should output logs to stdout with default log level WARN' do
+      expect(Logger).to receive(:new).with(STDERR, level: Logger::WARN).and_call_original
       foo.logger
     end
 
@@ -33,7 +28,7 @@ describe OpsManager::Logging do
       before { ENV['DEBUG']= 'true' }
 
       it 'should log in INFO level' do
-        expect_any_instance_of(Logger).to receive(:level=).with(Logger::INFO)
+      expect(Logger).to receive(:new).with(STDERR, level: Logger::INFO).and_call_original
         foo.logger
       end